requirements.txt
#devops
black==22.3.0
pytest==7.1.3
pytest-cov==4.0.0
requests==2.26.0
#rust based linter
ruff==0.0.284
databricks-sql-connector==3.5.0
python-dotenv==1.0.1